bugfix: always set `self.readonly_fields`, also when not set
This is different for `self.fields`, which the link fields should not be
appended to when it is not set, because when `self.fields` is empty it
means to show all fields. But the fields should always be added to
`self.readonly_fields` because otherwise they won't show up in the
change view.